requirements.txt
langchain==0.3.23
langchain-openai==0.1.1
langchain-core==0.3.52
langchain-text-splitters==0.3.8
langgraph==0.3.30
langsmith==0.3.31
streamlit==1.44.1
python-dotenv==1.1.0
faiss-cpu==1.8.0
requests==2.32.3
bs4==0.0.1
beautifulsoup4==4.12.3
openai==1.32.0
duckduckgo-search==4.5.0